Move output_settings to gui commmon in mantidqtinterfaces
